+++ graphql-native/src/main/java/io/ballerina/stdlib/graphql/runtime/engine/ResponseGenerator.java
@@ -25,6 +25,7 @@ import io.ballerina.runtime.api.creators.ValueCreator;
 import io.ballerina.runtime.api.types.ArrayType;
 import io.ballerina.runtime.api.utils.StringUtils;
 import io.ballerina.runtime.api.values.BArray;
+import io.ballerina.runtime.api.values.BError;
 import io.ballerina.runtime.api.values.BMap;
 import io.ballerina.runtime.api.values.BObject;
 import io.ballerina.runtime.api.values.BString;
@@ -36,6 +37,7 @@ import java.util.List;
 
 import static io.ballerina.stdlib.graphql.runtime.engine.Engine.executeResource;
 import static io.ballerina.stdlib.graphql.runtime.engine.Engine.getArgumentsFromField;
+import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.ERRORS_FIELD;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.IS_FRAGMENT_FIELD;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.KEY;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.NAME_FIELD;
@@ -44,8 +46,10 @@ import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.ON_TYPE_FIE
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.SELECTIONS_FIELD;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.copyAndUpdateResourcePathsList;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.createDataRecord;
+import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.getErrorDetailRecord;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.getNameFromRecordTypeMap;
 import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.isScalarType;
+import static io.ballerina.stdlib.graphql.runtime.engine.EngineUtils.updatePathSegments;
 
 /**
  * Used to generate the response for a GraphQL request in Ballerina.
@@ -54,32 +58,41 @@ public class ResponseGenerator {
     private ResponseGenerator() {
     }
 
-    static void getDataFromResult(Environment environment, BObject visitor, BObject node, Object result,
-                                  BMap<BString, Object> data, CallbackHandler callbackHandler) {
-        if (result instanceof BValue) {
+    static void populateResponse(Environment environment, BObject visitor, BObject node, Object result,
+                                 BMap<BString, Object> data, List<Object> pathSegments,
+                                 CallbackHandler callbackHandler) {
+        if (result instanceof BError) {
+            BError bError = (BError) result;
+            appendErrorToVisitor(bError, visitor, node, pathSegments);
+        } else if (result instanceof BValue) {
             int tag = ((BValue) result).getType().getTag();
             if (tag < TypeTags.JSON_TAG) {
                 data.put(node.getStringValue(NAME_FIELD), result);
             } else if (tag == TypeTags.RECORD_TYPE_TAG) {
-                getDataFromRecord(environment, visitor, node, (BMap<BString, Object>) result, data, callbackHandler);
+                getDataFromRecord(environment, visitor, node, (BMap<BString, Object>) result, data, pathSegments,
+                                  callbackHandler);
             } else if (tag == TypeTags.MAP_TAG) {
-                getDataFromMap(environment, visitor, node, (BMap<BString, Object>) result, data, callbackHandler);
+                getDataFromMap(environment, visitor, node, (BMap<BString, Object>) result, data, pathSegments,
+                               callbackHandler);
             } else if (tag == TypeTags.ARRAY_TAG) {
-                getDataFromArray(environment, visitor, node, (BArray) result, data, callbackHandler);
+                getDataFromArray(environment, visitor, node, (BArray) result, data, pathSegments, callbackHandler);
             } else if (tag == TypeTags.TABLE_TAG) {
-                getDataFromTable(environment, visitor, node, (BTable) result, data, callbackHandler);
+                getDataFromTable(environment, visitor, node, (BTable) result, data, pathSegments, callbackHandler);
             } else if (tag == TypeTags.SERVICE_TAG) {
                 getDataFromService(environment, (BObject) result, visitor, node, data, new ArrayList<>(),
-                                   callbackHandler);
+                                   pathSegments, callbackHandler);
             } // Here, `else` should not be reached.
         } else {
             data.put(node.getStringValue(NAME_FIELD), result);
         }
     }
 
+
     static void getDataFromService(Environment environment, BObject service, BObject visitor, BObject node,
-                                   BMap<BString, Object> data, List<String> paths, CallbackHandler handler) {
-        ResourceCallback callback = new ResourceCallback(environment, visitor, node, createDataRecord(), handler);
+                                   BMap<BString, Object> data, List<String> paths, List<Object> pathSegments,
+                                   CallbackHandler handler) {
+        ResourceCallback callback =
+                new ResourceCallback(environment, visitor, node, createDataRecord(), handler, pathSegments);
         handler.addCallback(callback);
         BArray selections = node.getArrayValue(SELECTIONS_FIELD);
         BMap<BString, Object> subData = createDataRecord();
@@ -89,10 +102,12 @@ public class ResponseGenerator {
             BObject subNode = selection.getObjectValue(NODE_FIELD);
             if (isFragment) {
                 if (service.getType().getName().equals(subNode.getStringValue(ON_TYPE_FIELD).getValue())) {
-                    executeResourceForFragmentNodes(environment, service, visitor, subNode, subData, paths, handler);
+                    executeResourceForFragmentNodes(environment, service, visitor, subNode, subData, paths,
+                                                    pathSegments, handler);
                 }
             } else {
-                executeResourceWithPath(environment, visitor, subNode, service, subData, paths, handler);
+                executeResourceWithPath(environment, visitor, subNode, service, subData, paths, pathSegments,
+                                        handler, i);
             }
             data.put(node.getStringValue(NAME_FIELD), subData);
         }
@@ -100,7 +115,8 @@ public class ResponseGenerator {
     }
 
     static void getDataFromRecord(Environment environment, BObject visitor, BObject node, BMap<BString, Object> record,
-                                  BMap<BString, Object> data, CallbackHandler callbackHandler) {
+                                  BMap<BString, Object> data, List<Object> pathSegments,
+                                  CallbackHandler callbackHandler) {
         BArray selections = node.getArrayValue(SELECTIONS_FIELD);
         BMap<BString, Object> subData = createDataRecord();
         for (int i = 0; i < selections.size(); i++) {
@@ -109,35 +125,42 @@ public class ResponseGenerator {
             BObject subNode = selection.getObjectValue(NODE_FIELD);
             if (isFragment) {
                 if (subNode.getStringValue(ON_TYPE_FIELD).getValue().equals(getNameFromRecordTypeMap(record))) {
-                    processFragmentNodes(environment, visitor, subNode, record, subData, callbackHandler);
+                    processFragmentNodes(environment, visitor, subNode, record, subData, pathSegments, callbackHandler);
                 }
             } else {
                 BString fieldName = subNode.getStringValue(NAME_FIELD);
                 Object fieldValue = record.get(fieldName);
-                getDataFromResult(environment, visitor, subNode, fieldValue, subData, callbackHandler);
+                List<Object> updatedPathSegments =
+                        updatePathSegments(pathSegments, node.getStringValue(NAME_FIELD).getValue());
+                populateResponse(environment, visitor, subNode, fieldValue, subData, updatedPathSegments,
+                                 callbackHandler);
             }
         }
         data.put(node.getStringValue(NAME_FIELD), subData);
     }
 
     static void getDataFromMap(Environment environment, BObject visitor, BObject node, BMap<BString, Object> map,
-                               BMap<BString, Object> data, CallbackHandler callbackHandler) {
+                               BMap<BString, Object> data, List<Object> pathSegments,
+                               CallbackHandler callbackHandler) {
         BMap<BString, Object> arguments = getArgumentsFromField(node);
         BString key = arguments.getStringValue(StringUtils.fromString(KEY));
         Object result = map.get(key);
-        getDataFromResult(environment, visitor, node, result, data, callbackHandler);
+        populateResponse(environment, visitor, node, result, data, pathSegments, callbackHandler);
     }
 
     static void getDataFromArray(Environment environment, BObject visitor, BObject node, BArray result,
-                                 BMap<BString, Object> data, CallbackHandler callbackHandler) {
+                                 BMap<BString, Object> data, List<Object> pathSegments,
+                                 CallbackHandler callbackHandler) {
         if (isScalarType(result.getElementType())) {
             data.put(node.getStringValue(NAME_FIELD), result);
         } else {
             BArray resultArray = ValueCreator.createArrayValue(getDataRecordArrayType());
             for (int i = 0; i < result.size(); i++) {
+                List<Object> updatedPathSegments = updatePathSegments(pathSegments, i);
                 Object resultElement = result.get(i);
                 BMap<BString, Object> subData = createDataRecord();
-                getDataFromResult(environment, visitor, node, resultElement, subData, callbackHandler);
+                populateResponse(environment, visitor, node, resultElement, subData, updatedPathSegments,
+                                 callbackHandler);
                 resultArray.append(subData.get(node.getStringValue(NAME_FIELD)));
             }
             data.put(node.getStringValue(NAME_FIELD), resultArray);
@@ -145,15 +168,17 @@ public class ResponseGenerator {
     }
 
     private static void getDataFromTable(Environment environment, BObject visitor, BObject node, BTable table,
-                                         BMap<BString, Object> data, CallbackHandler callbackHandler) {
+                                         BMap<BString, Object> data, List<Object> pathSegments,
+                                         CallbackHandler callbackHandler) {
         Object[] valueArray = table.values().toArray();
         ArrayType arrayType = TypeCreator.createArrayType(((BValue) valueArray[0]).getType());
         BArray valueBArray = ValueCreator.createArrayValue(valueArray, arrayType);
-        getDataFromResult(environment, visitor, node, valueBArray, data, callbackHandler);
+        populateResponse(environment, visitor, node, valueBArray, data, pathSegments, callbackHandler);
     }
 
     static void processFragmentNodes(Environment environment, BObject visitor, BObject node,
                                      BMap<BString, Object> record, BMap<BString, Object> data,
+                                     List<Object> pathSegments,
                                      CallbackHandler callbackHandler) {
         BArray selections = node.getArrayValue(SELECTIONS_FIELD);
         for (int i = 0; i < selections.size(); i++) {
@@ -161,40 +186,51 @@ public class ResponseGenerator {
             boolean isFragment = selection.getBooleanValue(IS_FRAGMENT_FIELD);
             BObject subNode = selection.getObjectValue(NODE_FIELD);
             if (isFragment) {
-                processFragmentNodes(environment, visitor, subNode, record, data, callbackHandler);
+                processFragmentNodes(environment, visitor, subNode, record, data, pathSegments, callbackHandler);
             } else {
                 BString fieldName = subNode.getStringValue(NAME_FIELD);
                 Object fieldValue = record.get(fieldName);
-                getDataFromResult(environment, visitor, subNode, fieldValue, data, callbackHandler);
+                List<Object> updatedPathSegments =
+                        updatePathSegments(pathSegments, node.getStringValue(NAME_FIELD).getValue());
+                populateResponse(environment, visitor, subNode, fieldValue, data, updatedPathSegments,
+                                 callbackHandler);
             }
         }
     }
 
     private static void executeResourceForFragmentNodes(Environment environment, BObject service, BObject visitor,
                                                         BObject node, BMap<BString, Object> data, List<String> paths,
-                                                        CallbackHandler callbackHandler) {
+                                                        List<Object> pathSegments, CallbackHandler callbackHandler) {
         BArray selections = node.getArrayValue(SELECTIONS_FIELD);
         for (int i = 0; i < selections.size(); i++) {
             BMap<BString, Object> selection = (BMap<BString, Object>) selections.get(i);
             boolean isFragment = selection.getBooleanValue(IS_FRAGMENT_FIELD);
             BObject subNode = selection.getObjectValue(NODE_FIELD);
             if (isFragment) {
-                executeResourceForFragmentNodes(environment, service, visitor, subNode, data, paths, callbackHandler);
+                executeResourceForFragmentNodes(environment, service, visitor, subNode, data, paths, pathSegments,
+                                                callbackHandler);
             } else {
-                executeResourceWithPath(environment, visitor, subNode, service, data, paths, callbackHandler);
+                executeResourceWithPath(environment, visitor, subNode, service, data, paths, pathSegments,
+                                        callbackHandler, i);
             }
         }
     }
 
     private static void executeResourceWithPath(Environment environment, BObject visitor, BObject node, BObject service,
                                                 BMap<BString, Object> data, List<String> paths,
-                                                CallbackHandler callbackHandler) {
+                                                List<Object> pathSegments, CallbackHandler callbackHandler, int index) {
         List<String> updatedPaths = copyAndUpdateResourcePathsList(paths, node);
-        executeResource(environment, service, visitor, node, data, updatedPaths, callbackHandler);
+        List<Object> updatedPathSegments = updatePathSegments(pathSegments, node.getStringValue(NAME_FIELD).getValue());
+        executeResource(environment, service, visitor, node, data, updatedPaths, updatedPathSegments, callbackHandler);
     }
 
     private static ArrayType getDataRecordArrayType() {
         BMap<BString, Object> data = createDataRecord();
         return TypeCreator.createArrayType(data.getType());
+    }
+
+    static void appendErrorToVisitor(BError bError, BObject visitor, BObject node, List<Object> pathSegments) {
+        BArray errors = visitor.getArrayValue(ERRORS_FIELD);
+        errors.append(getErrorDetailRecord(bError, node, pathSegments));
     }
 }