+++ changelog.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - [[#1190] Make GraphQL Resource Execution Non-Blocking](https://github.com/ballerina-platform/ballerina-standard-library/issues/1190)
 - [[#1218] Set `BAD_REQUEST` status code for responses with document validation errors](https://github.com/ballerina-platform/ballerina-standard-library/issues/1218)
+- [[#1507] Add Path Entry to the Error Detail](https://github.com/ballerina-platform/ballerina-standard-library/issues/1507)
 
 ### Fixed
 - [[#1447] Fix Returning Empty Values when Fragments Inside Fragments Querying Service Object](https://github.com/ballerina-platform/ballerina-standard-library/issues/1447)
@@ -19,6 +20,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - [[#1508] Fix Incorrectly Validating Invalid Operation Names when a Single Operation is Present in the Document](https://github.com/ballerina-platform/ballerina-standard-library/issues/1508)
 - [[#1505] Fix Field Order Not Maintained in the GraphQL Response](https://github.com/ballerina-platform/ballerina-standard-library/issues/1505)
 - [[#1526] Fix Parser Skipping Some Characters from Comments](https://github.com/ballerina-platform/ballerina-standard-library/issues/1526)
+- [[#1576] Fix Incorrect Response Order in GraphQL Response](https://github.com/ballerina-platform/ballerina-standard-library/issues/1576)
 
 ## [0.2.0-beta.1] - 2021-06-02